chore: rename path to /burn-dashboard
