+++ apps/web/src/views/BurnDashboard/components/StatsCard.tsx
