+++ apps/web/src/views/BurnDashboard/components/burns/WeeklyBurnStackedChart.tsx
@@ -6,9 +6,9 @@ import groupBy from 'lodash/groupBy'
 import { useCallback, useMemo, useState } from 'react'
 import { Bar, BarChart, ResponsiveContainer, Tooltip, TooltipProps, XAxis, YAxis } from 'recharts'
 import { formatAmount } from 'utils/formatInfoNumbers'
-import { CHART_COLORS } from 'views/CakeDashboard/constants'
-import { useBurnStats } from 'views/CakeDashboard/hooks/useBurnStats'
-import { getBurnInfoPrecision } from 'views/CakeDashboard/utils'
+import { CHART_COLORS } from 'views/BurnDashboard/constants'
+import { useBurnStats } from 'views/BurnDashboard/hooks/useBurnStats'
+import { getBurnInfoPrecision } from 'views/BurnDashboard/utils'
 import { StatsCard, StatsCardHeader } from '../StatsCard'
 import { TooltipCard } from '../styles'
 import { TabMenu } from '../TabMenu'
