+++ apps/web/src/views/BurnDashboard/components/emissions/EmissionsLastSevenDaysCard.tsx
@@ -1,8 +1,8 @@
 import { useTranslation } from '@pancakeswap/localization'
 import { CardProps, Text } from '@pancakeswap/uikit'
 import { formatAmount } from 'utils/formatInfoNumbers'
-import { useBurnStats } from 'views/CakeDashboard/hooks/useBurnStats'
-import { getBurnInfoPrecision } from 'views/CakeDashboard/utils'
+import { useBurnStats } from 'views/BurnDashboard/hooks/useBurnStats'
+import { getBurnInfoPrecision } from 'views/BurnDashboard/utils'
 import { StatsCard, StatsCardHeader } from '../StatsCard'
 
 export const EmissionsLastSevenDaysCard = (props: CardProps) => {
