+++ apps/web/src/views/BurnDashboard/components/emissions/EmissionsThirtyDaysCard.tsx
