+++ apps/web/src/views/BurnDashboard/components/emissions/WeeklyEmissionsStackedBarChart.tsx
@@ -7,9 +7,9 @@ import groupBy from 'lodash/groupBy'
 import { useMemo, useState } from 'react'
 import { Bar, BarChart, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts'
 import { formatAmount } from 'utils/formatInfoNumbers'
-import { CHART_COLORS_ALTERNATE } from 'views/CakeDashboard/constants'
-import { useBurnStats } from 'views/CakeDashboard/hooks/useBurnStats'
-import { getBurnInfoPrecision } from 'views/CakeDashboard/utils'
+import { CHART_COLORS_ALTERNATE } from 'views/BurnDashboard/constants'
+import { useBurnStats } from 'views/BurnDashboard/hooks/useBurnStats'
+import { getBurnInfoPrecision } from 'views/BurnDashboard/utils'
 import { StatsCard, StatsCardHeader } from '../StatsCard'
 import { TooltipCard } from '../styles'
 import { TabMenu } from '../TabMenu'
