+++ apps/web/src/views/BurnDashboard/components/general/SupplyPieChart.tsx
@@ -4,9 +4,9 @@ import { LightGreyCard } from 'components/Card'
 import { Cell, Pie, PieChart, ResponsiveContainer, Tooltip, TooltipProps } from 'recharts'
 import styled from 'styled-components'
 import { formatAmount } from 'utils/formatInfoNumbers'
-import { PEAK_SUPPLY } from 'views/CakeDashboard/constants'
-import { useBurnStats } from 'views/CakeDashboard/hooks/useBurnStats'
-import { getBurnInfoPrecision } from 'views/CakeDashboard/utils'
+import { PEAK_SUPPLY } from 'views/BurnDashboard/constants'
+import { useBurnStats } from 'views/BurnDashboard/hooks/useBurnStats'
+import { getBurnInfoPrecision } from 'views/BurnDashboard/utils'
 import { StatsCard, StatsCardHeader } from '../StatsCard'
 import { TooltipCard } from '../styles'
