+++ apps/web/src/views/BurnDashboard/components/styles.tsx
