+++ apps/web/src/views/BurnDashboard/constants.ts
