+++ apps/web/src/views/BurnDashboard/hooks/useBurnStats.ts
