+++ apps/web/src/views/BurnDashboard/types.ts
